+++ x/gov/common_test.go
@@ -4,9 +4,6 @@ import (
 	"bytes"
 	"log"
 	"sort"
-	"testing"
-
-	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
@@ -77,23 +74,3 @@ var (
 		ed25519.GenPrivKey().PubKey(),
 	}
 )
-
-func createValidators(t *testing.T, stakingHandler sdk.Handler, ctx sdk.Context, addrs []sdk.ValAddress, powerAmt []int64) {
-	require.True(t, len(addrs) <= len(pubkeys), "Not enough pubkeys specified at top of file.")
-
-	for i := 0; i < len(addrs); i++ {
-		valTokens := sdk.TokensFromConsensusPower(powerAmt[i], sdk.DefaultPowerReduction)
-		valCreateMsg, err := stakingtypes.NewMsgCreateValidator(
-			addrs[i], pubkeys[i], sdk.NewCoin(sdk.DefaultBondDenom, valTokens),
-			TestDescription, TestCommissionRates, sdk.OneInt(),
-		)
-		require.NoError(t, err)
-		handleAndCheck(t, stakingHandler, ctx, valCreateMsg)
-	}
-}
-
-func handleAndCheck(t *testing.T, h sdk.Handler, ctx sdk.Context, msg sdk.Msg) {
-	res, err := h(ctx, msg)
-	require.NoError(t, err)
-	require.NotNil(t, res)
-}